# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Drachenboot Manager API",
        "version": "2.5.0",
        "description": "API documentation for the Drachenboot Manager application with multi-team support."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/teams/{id}/api-keys": {
            "get": {
                "summary": "List API Keys",
                "description": "Lists all API keys for a team. Requires CAPTAIN role.",
                "tags": [
                    "API Keys"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of API keys",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApiKey"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create API Key",
                "description": "Creates a new API Key. Requires CAPTAIN role.",
                "tags": [
                    "API Keys"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "API Key created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "key": {
                                            "type": "string"
                                        },
                                        "record": {
                                            "$ref": "#/components/schemas/ApiKey"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/teams/{id}": {
            "put": {
                "summary": "Update Team",
                "description": "Updates a team's name. Requires CAPTAIN role.",
                "tags": [
                    "Teams"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Team ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "New team name"
                                    },
                                    "website": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "icon": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "primaryColor": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "showProRing": {
                                        "type": "boolean"
                                    },
                                    "showProBadge": {
                                        "type": "boolean"
                                    },
                                    "showWatermark": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only Captains can update the team",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update team",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Team",
                "description": "Deletes a team (Cascade). Requires CAPTAIN role.",
                "tags": [
                    "Teams"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Team ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete team",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "details": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Team": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique team identifier (CUID)"
                    },
                    "name": {
                        "type": "string",
                        "description": "Team name"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Team creation timestamp"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Last update timestamp"
                    },
                    "primaryColor": {
                        "type": "string",
                        "nullable": true
                    },
                    "showProRing": {
                        "type": "boolean"
                    },
                    "showProBadge": {
                        "type": "boolean"
                    },
                    "showWatermark": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
            "ApiKey": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "lastUsed": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message"
                    }
                },
                "required": [
                    "error"
                ]
            }
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "authjs.session-token",
                "description": "Session-based authentication using NextAuth.js. Requires active login session."
            }
        }
    },
    "security": [
        {
            "sessionAuth": []
        }
    ],
...
}